refactor(P7): extract ReferenceStatistics and ValidationResult to validation/types.py
Move the two public dataclasses out of validator.py into a dedicated
types module so callers can import lightweight type definitions without
pulling in scipy/numpy. Update __init__.py to export all three public
names; validator.py re-imports from types.py for zero breaking changes.
