+++ synthed/validation/__init__.py
@@ -1,3 +1,4 @@
+from .types import ReferenceStatistics, ValidationResult
 from .validator import SyntheticDataValidator
 
-__all__ = ["SyntheticDataValidator"]
+__all__ = ["SyntheticDataValidator", "ReferenceStatistics", "ValidationResult"]
